+++ src/main/java/com/alekseytyan/service/api/LorryService.java
@@ -6,5 +6,5 @@ import com.alekseytyan.entity.Lorry;
 import java.util.List;
 
 public interface LorryService extends AbstractService<Lorry, LorryDTO, String> {
-    List<LorryDTO> findSuitableLorries(Long weight);
+    List<LorryDTO> findSuitableLorries(int weight);
 }
